move fw2 product-specific code to its own file; more cleanup
